Re-encode if a website is not in UTF-8
It re-encodes in a particularly inefficient way: by parsing, searching
for the meta charset tag, re-encoding the raw bytes using that encoding,
and then re-parsing the re-encoded string.

Hyperfine (https://github.com/sharkdp/hyperfine) shows that this is 36%
slower than the previous version, but both versions complete nearly
instantaneously, so I'm OK with it.