Re-encode the string *before* it's parsed into HTML
By searching for the `<meta>` charset tag with a regular expression,
rather than a full HTML parser, it's more likely that we might not find
it even if it's there. However, it does save a lot of passing around
huge u8 buffers, and saves a full re-parse of the HTML.

Based on some quick benchmarks, this new approach is about 30% faster
than the old one.