Finders for the same node are shown together
For example, if the finders are:

* `div attr{class}` and
* `div {text}`

Then previously, it would show all of the `class`es, then all of the `text`s.

Now it will show the information for a given node next to each other. So
it will show each div's class, then that same div's text, then the next
div's class and text, and so on.