Align the pretty-printed text inside <script> tags
This behaves like the `String#squish` method in Rails: it detects the
indentation of the first line, then removes that amount of indentation
from each line after that. So if the first line is indented 4 spaces,
and the second one is indented 6 spaces, then the second one will be
indented 2 spaces more than the first.

If a line is indented *less* than the first line, like so:

      1
    2

...then it will be aligned with the first line:

    1
    2

In the final output, each line will be indented to its correct depth in
the tree, plus some amount of indentation will be added to each line
depending on its indentation relative to the first line.

For example, given this input:

```html
<body>
<div>
    <div>
    <script>
        var foo = "foo";
            var bar = "bar";
    function x(){
          return true;
        }
    </script>
    </div>
</div>
</body>
```

...then this is the output:

```html
<body>
  <div>
    <div>
      <script>
        var foo = "foo";
            var bar = "bar";
        function x(){
          return true;
        }
      </script>
    </div>
  </div>
</body>
```